Fix: Added social media model to model section
