--- packages/ui/src/cards/SocialMediaProfileCard/index.ts
@@ -1,1 +1,0 @@
-export * from "./SocialMediaProfileCard";
