+++ packages/ui/src/modals/SocialMediaModel/index.ts
@@ -1,0 +1,1 @@
+export * from "./SocialMediaModel";
