+++ packages/ui/src/modals/index.ts
@@ -2,3 +2,4 @@ export * from "./CongratulationsModal";
 export * from "./ProjectInfoModal";
 export * from "./RoleModalContainer";
 export * from "./SkillsModalContainer";
+export * from "./SocialMediaModel";
